Create Minimum Absolute Difference In BST
This solution defines a class Solution that calculates the absolute difference between adjacent nodes in a binary search tree (BST). Here's a breakdown:

The class Solution has an instance variable a initialized to Integer.MAX_VALUE, which will be used to store the minimum absolute difference found during traversal of the BST.

There's a method sol(Node root, int[] last) which is a recursive helper function to traverse the BST in an in-order manner (left subtree, root, right subtree). During traversal, it updates the minimum absolute difference (a) by comparing the absolute difference between the current node's data and the data of the last visited node (stored in last[0]). If last[0] is not -1, meaning it's not the first node visited, the absolute difference is calculated and compared to the current minimum (a).

The method absolute_diff(Node root) initializes the last array with -1 and then calls the sol() method to traverse the BST starting from the root. After traversal, it returns the minimum absolute difference (a).